To analise the correlation coeficent
This program is to analize the correlation coeficient between the heart rate measure by the Chip and the heart rate measure by Polar H10, which is the correct one. 
You have to run this sql in the supabase: select * from heart_rate_comparison_raw and then export this data into a csv file, put on the same directory that you run this program, and it will analise the results.